Trekk ut brevtyper og annen opprydding
