+++ src/frontend/komponenter/Behandling/Brev/BrevTyper.ts
@@ -1,0 +1,48 @@
+export type ValgtFelt = { [valgFeltKategori: string]: Valgmulighet };
+
+export interface BrevStruktur {
+    dokument: DokumentMal;
+    flettefelter: AlleFlettefelter;
+}
+export interface DokumentMal {
+    delmalerSortert: Delmal[];
+}
+
+export interface AlleFlettefelter {
+    flettefeltReferanse: Flettefelt[];
+}
+interface Flettefelt {
+    felt: string;
+    _id: string;
+}
+
+export interface Flettefeltreferanse {
+    _ref: string;
+}
+
+export interface FlettefeltMedVerdi extends Flettefeltreferanse {
+    verdi: string | null;
+}
+
+export interface Valgmulighet {
+    flettefelter: Flettefelter[];
+    valgmulighet: string;
+    visningsnavnValgmulighet: string;
+}
+
+export interface Flettefelter {
+    flettefelt: Flettefeltreferanse[];
+}
+export interface ValgFelt {
+    valgMuligheter: Valgmulighet[];
+    valgfeltVisningsnavn: string;
+    valgFeltApiNavn: string;
+}
+
+export interface Delmal {
+    delmalApiNavn: string;
+    delmalNavn: string;
+    delmalValgfelt: ValgFelt[];
+    delmalFlettefelter: Flettefelter[]; // referanse til flettefelt
+    mappe: string;
+}
