+++ src/frontend/komponenter/Behandling/Brev/BrevUtils.ts
@@ -1,4 +1,4 @@
-import { BrevStruktur } from './Brev';
+import { BrevStruktur } from './BrevTyper';
 
 export const finnFlettefeltNavnFraRef = (dokument: BrevStruktur, ref: string) => {
     return dokument.flettefelter.flettefeltReferanse.find((felt) => felt._id === ref)!.felt;
